// dmgcc/10_Landing_Page/src/saga/Types/aboutus.types.ts
export enum AboutUsActionTypes {
    //cost centre master
    GET_ABOUT_US_REQUEST = 'GET_ABOUT_US_REQUEST',
    GET_ABOUT_US_SUCCESS = 'GET_ABOUT_US_SUCCESS',
    GET_ABOUT_US_ERROR = 'GET_ABOUT_US_ERROR',
    GET_SERVICES_REQUEST = 'GET_SERVICES_REQUEST',
    GET_SERVICES_SUCCESS = 'GET_SERVICES_SUCCESS',
    GET_SERVICES_ERROR = 'GET_SERVICES_ERROR',
    GET_SERVICES_SUB_CONTENT_REQUEST = 'GET_SERVICES_SUB_CONTENT_REQUEST',
    GET_SERVICES_SUB_CONTENT_SUCCESS = 'GET_SERVICES_SUB_CONTENT_SUCCESS',
    GET_SERVICES_SUB_CONTENT_ERROR = 'GET_SERVICES_SUB_CONTENT_ERROR',
    GET_VISION_REQUEST = 'GET_VISION_REQUEST',
    GET_VISION_SUCCESS = 'GET_VISION_SUCCESS',
    GET_VISION_ERROR = 'GET_VISION_ERROR',
    GET_MISSION_REQUEST = 'GET_MISSION_REQUEST',
    GET_MISSION_SUCCESS = 'GET_MISSION_SUCCESS',
    GET_MISSION_ERROR = 'GET_MISSION_ERROR',
    GET_TESTIMONIAL_REQUEST = 'GET_TESTIMONIAL_REQUEST',
    GET_TESTIMONIAL_SUCCESS = 'GET_TESTIMONIAL_SUCCESS',
    GET_TESTIMONIAL_ERROR = 'GET_TESTIMONIAL_ERROR',
    GET_TEAM_MEMBERS_REQUEST = 'GET_TEAM_MEMBERS_REQUEST',
    GET_TEAM_MEMBERS_SUCCESS = 'GET_TEAM_MEMBERS_SUCCESS',
    GET_TEAM_MEMBERS_ERROR = 'GET_TEAM_MEMBERS_ERROR',
    GET_MISSION_VISION_REQUEST = 'GET_MISSION_VISION_REQUEST',
    GET_MISSION_VISION_SUCCESS = 'GET_MISSION_VISION_SUCCESS',
    GET_MISSION_VISION_ERROR = 'GET_MISSION_VISION_ERROR',
}

export interface AboutUsState {
    loading: boolean,
    items: {},
    errors:{},
    resAboutViewData?:any,
    resServicesData?:any,
    resServicesSubContentData?:any,
    resVisionData?:any,
    resMissionData?:any,
    resTestimonialData?:any,
    resTeamMembersData?:any,
    resMissionVisionData?:any,
    resVisionDetails?:any
}